changes in ItemProduto

+++ startApi/src/main/java/com/apirest/startApi/dto/itemProduto/DadosCadastroItemProdutoDto.java
@@ -10,7 +10,7 @@ public record DadosCadastroItemProdutoDto(
     @NotNull
     Produto produto,
     @NotNull @Positive
-    Double quantidade
+    Integer quantidade
 
 ) {
